Disable red zone
The "red zone" is part of the x86_64 ABI that allows userspace
programs to use a small chunk of stack below what they've officially
claimed with ESP. This is an optimization intended to allow small leaf
functions to avoid setting up a full stack frame.

In Kernel mode, the CPU's interrupt hardware isn't aware of this red
zone, and interrupts will clobber it. Before we can enable
interrupts (and thus also fault handling), we must stop using the red
zone and force functions to properly claim their stack regions.